Renamed files to conform to non pre z character.
