+++ src/cli.ts
@@ -1,17 +1,17 @@
 #!/usr/bin/env node
 
 import * as yargs from 'yargs';
-import { ZEsLint } from './zes-lint/zes-lint.class';
+import { ZEsLint } from './es-lint/es-lint.class';
 import { ZFileLint } from './file-lint/file-lint.class';
-import { ZHtmlHint } from './zhtml-hint/zhtml-hint.class';
-import { ZJsonLint } from './zjson-lint/zjson-lint.class';
+import { ZHtmlHint } from './html-hint/html-hint.class';
+import { ZJsonLint } from './json-lint/json-lint.class';
 import { ZConfigJsonParser } from './lint-config/config-json-parser.class';
 import { ZConfigNullReader } from './lint-config/config-null-reader.class';
 import { ZConfigReader } from './lint-config/config-reader.class';
-import { IZLintArgs } from './zlint/zlint-args.interface';
-import { ZLint } from './zlint/zlint.class';
+import { IZLintArgs } from './lint/lint-args.interface';
+import { ZLint } from './lint/lint.class';
 import { ZSassLint } from './zsass-lint/zsass-lint.class';
-import { ZTsLint } from './zts-lint/zts-lint.class';
+import { ZTsLint } from './ts-lint/ts-lint.class';
 import { ZYamlLint } from './zyaml-lint/zyaml-lint.class';
 
 const args: IZLintArgs = yargs.usage('$0 [options]').alias('c', 'config').describe('c', 'Optional config file to use.').string('c').help().parse() as any;
