+++ src/es-lint/es-lint.class.ts
@@ -1,6 +1,6 @@
 import chalk from 'chalk';
 import { CLIEngine } from 'eslint';
-import { IZLinter } from '../zlint/zlinter.interface';
+import { IZLinter } from '../lint/linter.interface';
 
 /**
  * Represents an object that can be used to perform eslint on javascript files.
