+++ src/html-hint/html-hint.class.ts
