+++ src/index.ts
@@ -1,17 +1,17 @@
-export * from './zes-lint/zes-lint.class';
+export * from './es-lint/es-lint.class';
 export * from './file-lint/content-linter.interface';
 export * from './file-lint/file-lint.class';
-export * from './zhtml-hint/zhtml-hint.class';
-export * from './zjson-lint/zjson-lint.class';
+export * from './html-hint/html-hint.class';
+export * from './json-lint/json-lint.class';
 export * from './lint-config/config-json-parser.class';
 export * from './lint-config/config-parser.interface';
 export * from './lint-config/config-reader.class';
 export * from './lint-config/config-reader.interface';
-export * from './zlint/zlint-args.interface';
-export * from './zlint/zlint-options.interface';
-export * from './zlint/zlint.class';
-export * from './zlint/zlinter.interface';
-export * from './zlint/zsilent-linter.class';
+export * from './lint/lint-args.interface';
+export * from './lint/lint-options.interface';
+export * from './lint/lint.class';
+export * from './lint/linter.interface';
+export * from './lint/silent-linter.class';
 export * from './zsass-lint/zsass-lint.class';
-export * from './zts-lint/zts-lint.class';
+export * from './ts-lint/ts-lint.class';
 export * from './zyaml-lint/zyaml-lint.class';
