+++ src/lint/lint-options.interface.ts
