+++ src/lint/linter.interface.ts
