+++ src/ts-lint/ts-lint.class.ts
